% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckAPIs.R
\name{CheckAPI}
\alias{CheckAPI}
\title{Check the Availability of APIs.}
\usage{
CheckAPI(
  database = c("SRA/ENA", "GEO", "PanglanDB", "UCSC Cell Browser", "Zenodo", "CELLxGENE",
    "Human Cell Atlas")
)
}
\arguments{
\item{database}{Vector of databases for API checking. Default: c("SRA/ENA", "GEO", "PanglanDB", "UCSC Cell Browser", "Zenodo", "CELLxGENE", "Human Cell Atlas").}
}
\value{
NULL.
}
\description{
Check the Availability of APIs.
}
\examples{
\donttest{
# check all databases
CheckAPI()
}
}
