% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSCCellBrowser.R
\name{ExtractCBComposition}
\alias{ExtractCBComposition}
\title{Extract Cell Type Composition of UCSC Cell Browser Datasets.}
\usage{
ExtractCBComposition(meta = NULL, link = NULL, json.folder = NULL)
}
\arguments{
\item{meta}{Metadata used to load, can be obtained with \code{ExtractCBDatasets}.
Skip when \code{link} is not NULL. Default: NULL.}

\item{link}{Vector contains dataset/collection link(s), e.g. "https://cells.ucsc.edu/?ds=adult-ureter".
Skip when \code{meta} is not NULL. Default: NULL.}

\item{json.folder}{Folder contains datasets json files, same as \code{json.folder} of \code{ShowCBDatasets}.
Default: NULL (access json files online).}
}
\value{
Dataframe contains sample information and cell type composition.
}
\description{
Extract Cell Type Composition of UCSC Cell Browser Datasets.
}
\examples{
\dontrun{
# lazy mode, load datasets json files locally, need users to provide json folder
ucsc.cb.samples <- ShowCBDatasets(lazy = TRUE, json.folder = NULL, update = FALSE)
# cell number is between 1000 and 2000
hbb.sample.df <- ExtractCBDatasets(
  all.samples.df = ucsc.cb.samples, organ = c("brain", "blood"),
  organism = "Human (H. sapiens)", cell.num = c(1000, 2000)
)
hbb.sample.ct <- ExtractCBComposition(json.folder = getwd(), meta = hbb.sample.df)
# for given dataset
ureter.ct <- ExtractCBComposition(link = "https://cells.ucsc.edu/?ds=adult-ureter")
}
}
