% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ExtractGEOMeta}
\alias{ExtractGEOMeta}
\title{Extract Sample Metadata from GEO.}
\usage{
ExtractGEOMeta(
  acce,
  platform = NULL,
  down.supp = FALSE,
  supp.idx = 1,
  timeout = 3600,
  ...
)
}
\arguments{
\item{acce}{GEO accession number.}

\item{platform}{Platform information/field. Default: NULL (all platforms).}

\item{down.supp}{Logical value, whether to download supplementary files to extract sample metadata. If TRUE, always
download supplementary files. If FALSE, extract GEO sample metadata. Default: FALSE.}

\item{supp.idx}{The index of supplementary files to download. Default: 1.}

\item{timeout}{Timeout for \code{\link{download.file}}. Default: 3600.}

\item{...}{Parameters for \code{\link{getGEO}}.}
}
\value{
Dataframe contains all metadata of provided GEO accession number.
}
\description{
Extract Sample Metadata from GEO.
}
\examples{
\donttest{
# users may need to set the size of the connection buffer
# Sys.setenv("VROOM_CONNECTION_SIZE" = 131072 * 60)
# extract GEO sample metadata of specified platform
GSE200257.meta <- ExtractGEOMeta(acce = "GSE200257", platform = "GPL24676")
# extract sample metadata from supplementary files
GSE292261.meta.supp <- ExtractGEOMetaSupp(acce = "GSE292261", supp.idx = 4)
}
}
