% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ExtractGEOMetaSupp}
\alias{ExtractGEOMetaSupp}
\title{Extract Sample Metadata from Supplementary Files.}
\usage{
ExtractGEOMetaSupp(acce, timeout = 3600, supp.idx = 1)
}
\arguments{
\item{acce}{GEO accession number.}

\item{timeout}{Timeout for \code{\link{download.file}}. Default: 3600.}

\item{supp.idx}{The index of supplementary files to download. Default: 1.}
}
\value{
A dataframe.
}
\description{
Extract Sample Metadata from Supplementary Files.
}
