% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca.R
\name{ExtractHCACatalogs}
\alias{ExtractHCACatalogs}
\title{Show All Available Catalogs in Human Cell Atlas.}
\usage{
ExtractHCACatalogs()
}
\value{
Named vector of catalogs.
}
\description{
Show All Available Catalogs in Human Cell Atlas.
}
\examples{
\donttest{
# all available catalogs
all.hca.catalogs <- ExtractHCACatalogs()
}
}
