% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanglaoDB.R
\name{ExtractPanglaoDBComposition}
\alias{ExtractPanglaoDBComposition}
\title{Extract Cell Type Composition of PanglaoDB Datasets.}
\usage{
ExtractPanglaoDBComposition(
  meta = NULL,
  sra = NULL,
  srs = NULL,
  local.data = TRUE
)
}
\arguments{
\item{meta}{Metadata contains "SRA", "SRS", can be obtained with \code{ExtractPanglaoDBMeta}.
Skip when \code{sra} or \code{srs} is not NULL. Default: NULL.}

\item{sra}{The SRA identifier of the datasets. Skip when \code{meta} or \code{srs} is not NULL. Default: NULL.}

\item{srs}{The SRS identifier of the datasets. Skip when \code{meta} or \code{sra} is not NULL. Default: NULL.}

\item{local.data}{Logical value, whether to use local data (PanglaoDB is no longer maintained). Default: TRUE.}
}
\value{
Dataframe contains sample metadata, cluster, cell number and cell type information.
}
\description{
Extract Cell Type Composition of PanglaoDB Datasets.
}
\examples{
\donttest{
# get metadata
human.meta <- ExtractPanglaoDBMeta(
  species = "Homo sapiens",
  protocol = c("Smart-seq2", "10x chromium"),
  cell.num = c(1000, 2000)
)
# get composition
human.composition <- ExtractPanglaoDBComposition(
  meta = human.meta
)
# given dataset
human.composition <- ExtractPanglaoDBComposition(sra = "SRA598936")
# human.composition <- ExtractPanglaoDBComposition(srs = "SRS2428405")
# human.composition <- ExtractPanglaoDBComposition(sra = "SRA598936", srs = "SRS2428405")
}
}
