% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessRData.R
\name{LoadRData}
\alias{LoadRData}
\title{Load Objects, Count Matrix, and Metadata from RData.}
\usage{
LoadRData(
  rdata,
  accept.fmt = c("Seurat", "seurat", "SingleCellExperiment", "cell_data_set",
    "CellDataSet", "DESeqDataSet", "DGEList"),
  show.object = TRUE,
  return.obj = TRUE,
  slot = c("counts", "data", "scale.data")
)
}
\arguments{
\item{rdata}{Path to RData file.}

\item{accept.fmt}{Vector, the class of objects for loading. Default: c("Seurat", "seurat", "SingleCellExperiment",
"cell_data_set", "CellDataSet", "DESeqDataSet", "DGEList"). "Seurat" for Seurat v3, v4; "seurat" for Seurat v2.}

\item{show.object}{Logical value, whether to show the class of available objects. Default: TRUE.}

\item{return.obj}{Logical value, whether to load the available objects in \code{accept.fmt} to global environment. Default: TRUE.}

\item{slot}{Vector, the type of count matrix to pull. Default: c("counts", "data", "scale.data").
counts: raw, un-normalized counts, data: normalized data, scale.data: z-scored/variance-stabilized data.}
}
\value{
List contains count matrix and metadta or NULL.
}
\description{
Load Objects, Count Matrix, and Metadata from RData.
}
\examples{
\dontrun{
# download RData from GEO
ParseGEOProcessed(acce = "GSE244572", timeout = 360000, supp.idx = 1, out.folder = "/path/to/RData", file.ext = c("rdata", "rds", "h5ad", "loom"))
# with six assays: "RNA", "ADT", "nADT", "SCT", "integrated", "IADT"
multi.assyas.list <- LoadRData(
  rdata = "/path/to/GSE244572_RPE_CITESeq.RData",
  accept.fmt = c("Seurat", "seurat", "SingleCellExperiment", "cell_data_set", "CellDataSet", "DESeqDataSet", "DGEList"),
  slot = "counts"
)
# only one object in GSE244572_RPE_CITESeq.RData
names(multi.assyas.list)
# metadata
multi.assyas.list$obj$meta.data \%>\% head()
# assay names
names(multi.assyas.list$obj$count.mat)
# raw count matrix (slot)
multi.assyas.list$obj$count.mat$RNA$counts[1:5, 1:5]
}
}
