% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ParseGEOProcessed}
\alias{ParseGEOProcessed}
\title{Download Processed Objects from GEO.}
\usage{
ParseGEOProcessed(
  acce,
  timeout = 3600,
  supp.idx = 1,
  file.ext = c("rdata", "rds", "h5ad", "loom"),
  out.folder = NULL,
  return.seu = FALSE,
  merge = TRUE
)
}
\arguments{
\item{acce}{GEO accession number.}

\item{timeout}{Timeout for \code{\link{download.file}}. Default: 3600.}

\item{supp.idx}{The index of supplementary files to download. Default: 1.}

\item{file.ext}{The valid file extension for download (ignore case and gz suffix). When NULL, use all files. Default: c("rdata", "rds", "h5ad").}

\item{out.folder}{The output folder. Default: NULL (\code{acce} folder under current working directory).}

\item{return.seu}{Logical value, whether to load downloaded datasets to Seurat. Valid when rds in \code{file.ext} and all
datasets download successfully. Default: FALSE.}

\item{merge}{Logical value, whether to merge Seurat list when there are multiple rds files,
used when \code{return.seu} is TRUE. Default: FALSE.}
}
\value{
SeuratObject (\code{return.seu} is TRUE, rds in \code{file.ext}) or
NULL (\code{return.seu} is FALSE or rds not in \code{file.ext}).
}
\description{
Download Processed Objects from GEO.
}
\examples{
\dontrun{
# need users to provide the output folder
# suitable for rdata, rdata.gz, rds, rds.gz
# return SeuratObject when return.seu = TRUE
GSE285723.seu <- ParseGEOProcessed(
  acce = "GSE285723", supp.idx = 1,
  file.ext = c("rdata", "rds"), return.seu = TRUE,
  out.folder = "/path/to/outfoder"
)
geo.rdata.log <- ParseGEOProcessed(
  acce = "GSE311825", supp.idx = 4,
  file.ext = c("rdata", "rds"),
  out.folder = "/path/to/outfoder"
)
# suitable for h5ad, h5ad.gz
geo.h5ad.log <- ParseGEOProcessed(
  acce = "GSE311813", supp.idx = 1,
  file.ext = c("h5ad"),
  out.folder = "/path/to/outfoder"
)
# suitable for loom, loom.gz
geo.loom.log <- ParseGEOProcessed(
  acce = "GSE286325", supp.idx = 1,
  file.ext = c("loom"),
  out.folder = "/path/to/outfoder"
)
}
}
