% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca.R
\name{ShowHCAProjects}
\alias{ShowHCAProjects}
\title{Show All Available Projects in Human Cell Atlas.}
\usage{
ShowHCAProjects(catalog = NULL)
}
\arguments{
\item{catalog}{The catalog of the projects (one or multiple values). Different catalogs may share some projects.
All available catalogs can be accessed via \code{ExtractHCACatalogs}. Default: NULL (all catalogs, remove duplicated projects).}
}
\value{
Dataframe contains all available projects.
}
\description{
Show All Available Projects in Human Cell Atlas.
}
\examples{
\donttest{
# all available projects
all.hca.projects <- ShowHCAProjects()
}
}
