% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ParseGEOProcessed}
\alias{ParseGEOProcessed}
\title{Download Processed Objects from GEO.}
\usage{
ParseGEOProcessed(
  acce,
  timeout = 3600,
  supp.idx = 1,
  file.ext = c("rdata", "rds", "h5ad", "loom"),
  out.folder = NULL
)
}
\arguments{
\item{acce}{GEO accession number.}

\item{timeout}{Timeout for \code{\link{download.file}}. Default: 3600.}

\item{supp.idx}{The index of supplementary files to download. Default: 1.}

\item{file.ext}{The valid file extension for download (ignore case and gz suffix). When NULL, use all files. Default: c("rdata", "rds", "h5ad").}

\item{out.folder}{The output folder. Default: NULL (\code{acce} folder under current working directory).}
}
\value{
NULL.
}
\description{
Download Processed Objects from GEO.
}
\examples{
\dontrun{
# need users to provide the output folder
# suitable for rdata, rdata.gz, rds, rds.gz
geo.rds.log <- ParseGEOProcessed(
  acce = "GSE285723", supp.idx = 1,
  file.ext = c("rdata", "rds"),
  out.folder = "/path/to/outfoder"
)
geo.rdata.log <- ParseGEOProcessed(
  acce = "GSE311825", supp.idx = 4,
  file.ext = c("rdata", "rds"),
  out.folder = "/path/to/outfoder"
)
# suitable for h5ad, h5ad.gz
geo.h5ad.log <- ParseGEOProcessed(
  acce = "GSE311813", supp.idx = 1,
  file.ext = c("h5ad"),
  out.folder = "/path/to/outfoder"
)
# suitable for loom, loom.gz
geo.loom.log <- ParseGEOProcessed(
  acce = "GSE286325", supp.idx = 1,
  file.ext = c("loom"),
  out.folder = "/path/to/outfoder"
)
}
}
